-fixed space padding of charger status strings in hmi.c -added "IDLE" status string for charger idle state -fixed wrong type for chargePowerPrecise -fixed transition from pass-through mode to MPPT: this needs a MPPT PWM up in the first MPPT step or the system will oscillate between passthrough and MPPT. -added initialization of MPPT_direction_up to "up" = 0xFF.
git-svn-id: file:///home/dermeisterr/AVR%20Studio%20Projekte/SVN-Repository/trunk@86 c29d74f2-9f64-704e-acbf-26847afd32bc

+++ SLACC_1-1/src/measurement.c
@@ -68,9 +68,9 @@ void measure(void)
     measurements.temperature2.v = linearizeU16(&linListKty81210, measurements.temperature2.adc);
     
     // Compute values
-    uint32_t chargePowerPrecise = (uint32_t)measurements.batteryVoltage.v * (uint32_t)measurements.chargeCurrent.v;
+    uint64_t chargePowerPrecise = (uint32_t)measurements.batteryVoltage.v * (uint32_t)measurements.chargeCurrent.v;
     measurements.panelPower = (uint16_t)((uint32_t)measurements.panelVoltage.v * (uint32_t)measurements.panelCurrent.v / 10000UL);
-    measurements.chargePower = (uint16_t)(chargePowerPrecise / 10000UL);
+    measurements.chargePower = (uint16_t)(chargePowerPrecise / (uint64_t)10000);
     
     if (measurements.panelPower > measurements.chargePower)
         measurements.efficiency = (uint16_t)(chargePowerPrecise / measurements.panelPower);
